backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.6.0
pymilvus==2.3.3
httpx==0.27.2
pydantic==2.8.2
python-frontmatter==1.1.0
tenacity==9.0.0
PyJWT==2.8.0
bcrypt==4.1.2

